+++ src/main/java/ganymedes01/etfuturum/api/world/IGeneratingCheck.java
@@ -1,0 +1,6 @@
+package ganymedes01.etfuturum.api.world;
+
+public interface IGeneratingCheck {
+    /// Checks if this world is currently running generation code. If false, worldgen code is not running.
+    boolean etfuturum$isGenerating();
+}